+++ src/Simple.Logging.Console/LogLevelColors.cs
@@ -1,8 +1,0 @@
-namespace Simple.Logging.Console;
-
-public readonly record struct LogLevelColors(
-    ILogColor? BadgeForeground,
-    ILogColor? BadgeBackground,
-    ILogColor MessageColor,
-    ILogColor HighlightColor,
-    ILogColor AccentColor);